// client/src/types.ts
export interface Message {
  id: string;
  senderId: string;
  receiverId: string;
  content: string;
  imagePath?: string;
  timestamp: string;
  isRead: boolean;
}

export interface Conversation {
  id: string;
  participantId: string;
  participantName: string;
  participantType: string;
  participantImage?: string;
  lastMessage: string;
  lastMessageTime: string;
  unreadCount: number;
}